refactor: split Field.kt into U256.kt, FieldP.kt, ScalarN.kt
Each object now lives in its own file for easier navigation and review:

- U256.kt (284 lines): Raw 256-bit unsigned integer arithmetic with the
  file-level architecture documentation explaining representation choices
- FieldP.kt (360 lines): Field arithmetic modulo the secp256k1 prime p,
  including reduction, inversion, and square root
- ScalarN.kt (230 lines): Scalar arithmetic modulo the group order n,
  including wide reduction and Fermat inversion

No functional changes — pure file reorganization.
